fix(file): When File is filtered do not write its content to disk.
BREAKING CHANGE: This moves filter configurations from Collections to
File. This is required as we need to know if a File should be written
to disk, and Collections delegate their behavior depending on how the
File is configured.

This is an easy transition, mostly moving the filter configuration
from the Collection to File under the `filters` key.

Fixes #39.